Multitasking: Drive traditional threads from protothreads
In the previous architecture, there was a traditional thread called
`minitask-runner` that executed protothreads. As we migrate more
and more tasks into protothreads, it makes sense to invert this.

Now, the protothreads run in a loop after the kernel initializes. In
order to keep the old threads working, we yield to them whenever all
the protothreads have been run.