serial: Buffer logs before we start the serial writer
Before we start multitasking and the serial writer service the logs
we produced were not being captured. This change creates the
SerialWriter buffer on the first write attempt, and outputs the
captured logs when the writer service starts.

This change, combined with a larger buffer, allows us to get all the
logs from the boot process after everything starts up.