runtime: Load and execute BF code
The OS was unable to execute any code that wasn't compiled into the
kernel binary before. This commit introduces a BF runtime that can
read a BF file from disk and execute it.

The runtime doesn't support too many useful instructions, but it's
enough to play around with and demonstrate concepts.