scheduler: Allow tasks to be blocked
Previously, a task could be taken out of the scheduler by sleeping.
It would continue being scheduled after its sleep timer ran out. This
is useful for waiting a specific duration, but not very useful when
a task needs to be blocked until another tasks requires something.

This commit adds the following methods to block and unblock tasks.

- block()
- unblock()
- isBlocked()