io: Block serial writer when the write buffer is empty
When there is nothing to be written through the serial port, the
serial writer task should block itself to conserve CPU. This prevents
it from being scheduled regularly only to make itself sleep again.

When something is written to the buffer, the task is immediately
unblocked. On top of the increased efficiency, this might even
improve the output latency since the sleeps are now eliminated.