drivers: Really basic PCI driver
This commit adds a rudimentary PCI driver. Right now, all it can do
is test the availability of PCI, and to print a list of PCI devices
that were found.

There is no PCI device database yet, and nothing is done when a
device is found other than logging it in the serial output.